refactor active marker, allow to zoom and pan
