+++ pages/narratives/_book/_chapter/index.vue
@@ -4,24 +4,4 @@
     <card-list v-bind:cards="pages" />
   </div>
 </template>
-
-<script>
-import CardList from '~/components/card-list/CardList'
-import loadData from '~/lib/load-data'
-
-export default {
-  layout: 'list',
-  asyncData (context) {
-    return loadData(context, context.params)
-  },
-  components: {
-    CardList
-  },
-  data () {
-    return {
-      title: '',
-      pages: []
-    }
-  }
-}
-</script>
+<script src="./index-component.js"></script>
